refactor(server): explicitly cast resource version to integer
The wl_resource_create function takes version as an integer. It is not clear
why it is done this way as version numbers are declared as uint32_t in the
Wayland documentation [1].

For now just static cast to integer for this call.

[1] https://wayland.freedesktop.org/docs/html/apa.html#id-1.9.4.6

+++ server/wayland/client.cpp
@@ -103,7 +103,7 @@ wl_resource* Client::createResource(const wl_interface* interface, uint32_t vers
     if (!m_client) {
         return nullptr;
     }
-    return wl_resource_create(m_client, interface, version, id);
+    return wl_resource_create(m_client, interface, static_cast<int>(version), id);
 }
 
 wl_resource* Client::getResource(uint32_t id)
